requirements.txt
flask>=3.0.0
opencv-python>=4.6.0
pillow>=9.0.0
numpy>=1.20.0
git+https://github.com/THU-MIG/yolov10.git 